requirements.txt
azure-ai-formrecognizer==3.3.3
azure-core==1.30.0
python-dotenv==1.0.1
